Updated Terms and Definitions
This is now a standalone document that is included as a document object
in other Phase 4 documents. This means that edits to terms and
definitions in the standalone document automatically show up wherever
it’s included as a document object. This simplifies and centralizes the
management of Terms and Definitions.