Puncturing paper that "sheds light" (get it? optical?)
Puncturing means removing symbols. Isn’t that bad? Well, it can be, if
you don’t do it intelligently. This paper talks about what kind of
performance penalty one can expect to get with the particular type of
puncturing used in DVB-S2. This is background, more or less, to the
DVB-S2X puncturing.