Code of Conduct guide
Having a code of conduct is one thing. Enforcing it is another. Here's a book about how to use codes of conduct. Phase 4 Ground has a code of conduct that is outlined on the Open Research Institute website. This book is a resource for understanding how to use tools such as CoCs effectively.