Add files via upload
Note: the document title includes the acronym AREx (Amateur Radio Experiment). It's unclear whether this term is considered by ARISS-USA to be "owned by ARISS" - or not. It is not clear if ARISS will object to others using it outside ARISS. It has occurred in amateur radio literature in the past for projects unrelated to space. In other words, "Amateur Radio Experimentation" is not a unique term or acronym. No projects as yet using the term have risen in popular amateur media to the level where it becomes a brand. Unless and until we're told otherwise, AREx, for the purpose of this document, means the general practice of Amateur Radio Experimentation. 

The content of this document is the work of ORI volunteers working on AREx/ARISS/AMSAT, while AREx meetings were held in 2020. The AREx project was "rebooted" in late summer 2020. Despite the technical contributions in this document, and a written offer to share fundraising and provide a commitment of technical support, ARISS/AMSAT excluded ORI from participating at the "reboot". No reason was given outside of 1) an objection to having public design reviews for open source work and 2) using the Agile management style as opposed to Waterfall.

A difference in management style is a solvable root cause. The lack of interest in integrating organizations with different styles was disappointing, but this is not an uncommon challenge in collaboration. More worrisome, there seemed to be a profoundly USA-centric discomfort and inability to take advantage of open source products and processes.  Open source satellite work is free of ITAR. The work to free it from EAR is ongoing. These innovations provide enormous advantage and relief to amateur satellite organizations in the US, but as we can see in this case, there is a reluctance to take advantage of those developments. 

ORI board discussed this situation and was dismayed. ORI believes that publicly funded opportunities (NASA missions) should be accessible to the public. Organizations like ARISS that set themselves up to be gatekeepers for publicly funded institutions should be "opening the gate" as widely as possible. Shutting it to volunteers outside of narrowly defined groups that adhere to narrowly defined management styles, failing to take advantage of regulatory innovations, passing over superior amateur-developed technology, and sidelining large numbers of available volunteer hours - all of this  is bad for the Amateur Radio Satellite Service. 

Fortunately, this situation should be short term. Minds can be changed by good examples and clear regulatory wins for the Service. If you would like to help in this effort, please visit 

https://www.openresearch.institute/getting-started/

And let us know you want to help with regulatory (or technical) work.