propa.dll file from Jan King, to be used with his spreadsheet
The ITU Rain Model uses macros and it makes use of a data base, which computes rain, atmospheric and scintillation losses appropriate for all of the specified conditions in the model. This is all contained in a systems file called propa.dll. You put this file into your WINDOWS10 machine (this thing won't work on a Mac) operating folder using the procedures in the Word file (also attached).  Then reboot the machine; then install and open the Excel Link Model attached.  When you open this Excel as it is a .xlsm type file, it is suitable for using macros.  When you open the file it will prompt you by asking if you want to enable the macros.  Say "yes".