Preliminary Lab Results from Mike Parker et al
Mike Parker and David Chan measured the 10/24 GHz dual band feed in a lab at Rincon. Here are the preliminary results.

David writes: "this is an E-plane measurement, I’m waiting on a couple WR-42 waveguides to come in so I can make all the tests easily in one go."