A few sentences of intro, and GUI fail workaround
A few sentences of introduction.
There's a problem where on chococat we can't select "create project platform" this is a common enough error to where a lot of other people experiencing it have posted in xilinx and analog devices forums. Command line platform project creation is the way to get around this very annoying problem with Vitis.