Added working extraction code for Project-NiCd.
Working extraction code for Project-NiCd. This code reads the .csv files in the data folder (if it is placed in a directory with them) and utilizes the string signatures to obtain the exact section of the data that are for each cycle. Furthermore, it also generates a text printout of the <Messages> blob at the end of all of the .uba files also stored in the data folder. This code utilizes os, pandas, numpy and contextlib.