Updated extraction code for Project-NiCd
Edited and clarified statements in comments, as well as created more efficient dictionary assignment for conditioned data. This file is otherwise the same as rwout.ipynb, utilizing pandas, numpy, itertools and os, with manipulation of the .csv and .uba files. Potential issues may be that the conditioned data is being inaccurately truncated.