Updated extraction code for Project-NiCd (#55)
* Updated extraction code for Project-NiCd

Edited and clarified statements in comments, as well as created more efficient dictionary assignment for conditioned data. This file is otherwise the same as rwout.ipynb, utilizing pandas, numpy, itertools and os, with manipulation of the .csv and .uba files. Potential issues may be that the conditioned data is being inaccurately truncated.

* Delete Remote_Labs/Project_NiCd/rwout.ipynb

Replaced with ExtractionCode.ipynb

* Delete Remote_Labs/Project_NiCd/ExtractionCode.ipynb

Replacing with updated version with revisions

* Updated and revised extraction code

Edited and revised to utilize index numbers rather than column values, as that was causing issues in the readout of assigned files.

* Delete Remote_Labs/Project_NiCd/ExtractionCode.ipynb

Replacing with updated and expanded code

* Updated and Expanded Extraction Code

Added resulting .csv files from new code that utilizes dataframes and pandas to compare our test cycles directly. Otherwise, not different from previous extraction code.