Who is Afraid of SAR presentation slides
"Who is Afraid of SAR" was presented to DEFCON RF Village on 10 August 2024. SAR stands for Synthetic Aperture Radar, which is a technology controlled by ITAR and appears on the United States Munitions List. By taking advantage of the public domain carve-out, and by only making articles for educational or R&D purposes, we can design, build, and publish designs for SAR.