Correct noise bandwidth per mode
added a "hidden" column of the "real" noise bandwidth. The JT modes give an SNR and a bandwidth, but the bandwidth the SNR is calculated at is 2500 and not what's given per mode. We calculate the 1 Hz SNR per mode using 2500 Hz for the JT modes, and the "real" bandwidth for the legacy modes.