Made *sure* Doppler spread is Gaussian
Made sure that the Doppler spread model uses a Gaussian distribution. Added a calculation for the limits of the Doppler Spread and added an explanation that the vast majority of the energy does reflect from the facing side of Venus, and that the 31 Hz Doppler Spread is a distribution. We are modeling it as Gaussian for right now, but a better model might be a narrower Beta distribution, based on radar return data from 1965 Soviet surveys, and other papers that we can see in online searches.