PDF export working, pointing error cells combined
PDF export is working now. The SVG and the Doppler Spread worksheet graphics were not exporting correctly. This was fixed by writing an embedder for the SVG and the plotly graphics for the Doppler Spread. This does require kaleido, which installed in poetry environment for me when using the command `poetry add kaleido==0.2.1`

The pointing error cells have been combined. The tracking error has been dropped. The reported pointing error has been added to the visualization. The explanations are much better now.