Add files via upload
Added introductory and explanatory text that was developed for the March 2025 Inner Circle Newsletter article for the link budget. Added a section about the different sites that are involved or potentially can be involved. Multiple improvements in all sections up to Doppler were made in this check-in. 

In Pointing Loss Section of the Code:
Changed all occurrences of the divisor 5.66 to np.sqrt(12) (approximately 3.46)
Renamed the variable loss_at_166 to loss_at_sqrt12 for clarity
Updated the annotations to reference "beamwidth/√12" instead of "beamwidth/5.66"
Used the correct variables in the plotting functions

These changes ensure mathematical consistency between the formula pointing_loss = -12 * (error_angle / beamwidth)² and the derived thresholds. Now when a pointing error equals beamwidth/√12, it will correctly result in a 1dB loss.