Add files via upload
This is a significant update because it incorporates updated Doppler Spread modeling and calculations into the link budget. These are based on multiple sets of observations, but most recently and most notably those from @daniestevez in his blog post at https://destevez.net/2025/04/analysis-of-the-camras-venus-radar-experiment/ 

The markdown has been greatly expanded to explain double-shift phenomenon in radar returns, motivation for updating the model, and how updating it was achieved. Subsequent blocks have been updated to use the new results.