Add files via upload
improvements to the markdown text and to the code. Markdown text has multiple improvements for clarity and accuracy. For usage, one is now directed to the code, and that usage has been improved. Control over debug statements being printed has been incorporated and tested. Explanation of the doppler and doppler optimization algorithms has been included.