Add files via upload
Updated Zadoff-Chu code and markdown. Added brief explanation of what coherent and non-coherent modes are (an outstanding action item) and made it calculate the maximum possible integration time from Doppler Rate of Change and then calculate the coherence time from Doppler spread. The more conservative number wins as the coherence time available for the Zadoff-Chu sequence. The section proceeds to try to see if the selected rate works to close the link, and if it doesn't, then how many more non-coherent integrations would be needed to get over the specified link margin.